requirements.txt
Django>=4.2,<5.0
Pillow>=10.0
google-generativeai>=0.4.0